Remy did some cleanup but is afriad of touching drivebase
